requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiohttp>=3.9.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pypdf>=3.17.0
numpy>=1.24.0
python-dotenv>=1.0.0
